+++ Source/dsp/NonLinearEquationBase.h
@@ -17,8 +17,8 @@ class NonLinearEquationBase
 public:
 
     virtual ~NonLinearEquationBase() = default;
-    virtual float calculateCurrents(Eigen::MatrixXf voltages, int vn_index, int function_index) = 0;
-    virtual void calculateJacobian(Eigen::MatrixXf * jacobian, Eigen::MatrixXf voltages, int vn_index) = 0;
+    virtual float calculateCurrents(Eigen::MatrixXd voltages, int vn_index, int function_index) = 0;
+    virtual void calculateJacobian(Eigen::MatrixXd * jacobian, Eigen::MatrixXd voltages, int vn_index) = 0;
     int getNumberOfFunctions() { return numberOfFunctions; };
 protected:
